tests/Dataplace.Imersao.Core.Tests/Domain/Orcamentos/Enums/OrcamentoStatusEnumTest.cs
using Dataplace.Imersao.Core.Domain.Orcamentos.Enums;
using Xunit;

namespace Dataplace.Imersao.Core.Tests.Domain.Orcamentos.Enums
{
    public class OrcamentoStatusEnumTest
    {
        [Fact]
        [Trait("OrcamentoStatusEnum", "ToDataValue")]
        public void OrcamentoFechadoToDataValueTest()
        {

            // arrange act assert
            Assert.Equal("P", OrcamentoStatusEnum.Aberto.ToDataValue());
            Assert.Equal("F", OrcamentoStatusEnum.Fechado.ToDataValue());
            Assert.Equal("C", OrcamentoStatusEnum.Cancelado.ToDataValue());
        }

        [Fact]
        [Trait("OrcamentoStatusEnum", "ToOrcamentoStatusEnum")]
        public void OrcamentoAbertoToDataValueTest()
        {
            // arrange act assert
            Assert.Equal(OrcamentoStatusEnum.Aberto, "P".ToOrcamentoStatusEnum());
            Assert.Equal(OrcamentoStatusEnum.Fechado, "F".ToOrcamentoStatusEnum());
            Assert.Equal(OrcamentoStatusEnum.Cancelado, "C".ToOrcamentoStatusEnum());
            Assert.Equal(OrcamentoStatusEnum.Aberto, "X".ToOrcamentoStatusEnum());
            Assert.Equal(OrcamentoStatusEnum.Aberto, default(string).ToOrcamentoStatusEnum());
        }
    }
}
